chore: remove new axelarqueryclient, refactor to use existing cosmos package
