+++ packages/api/src/axelar-query/types.ts
@@ -1,4 +1,11 @@
 /* eslint-disable @typescript-eslint/no-duplicate-enum-values */
+import { QueryService as AxelarnetQS } from "@axelarjs/proto/axelar/axelarnet/v1beta1/service";
+import { QueryService as EvmQS } from "@axelarjs/proto/axelar/evm/v1beta1/service";
+import { QueryService as NexusQS } from "@axelarjs/proto/axelar/nexus/v1beta1/service";
+import { QueryService as TSSQS } from "@axelarjs/proto/axelar/tss/v1beta1/service";
+
+import type { QueryClient } from "@cosmjs/stargate";
+
 import type { GetFeesParams } from "..";
 
 export type EstimateGasFeeParams = GetFeesParams & {
@@ -99,3 +106,10 @@ export interface AxelarGMPResult {
 export interface AxelarGMPResponse {
   result: AxelarGMPResult;
 }
+export interface AxelarQueryService {
+  readonly evm: EvmQS;
+  readonly axelarnet: AxelarnetQS;
+  readonly nexus: NexusQS;
+  readonly tss: TSSQS;
+}
+export type AxelarQueryClientType = QueryClient & AxelarQueryService;
